monitor-manager/requirements.txt
psutil==5.9.6
Pillow
PyQt5==5.15.10
